+++ src/adt/def.jl
@@ -233,7 +233,7 @@ function iscanonical(psi::ADT; kwargs...)
 	return true
 end
 
-function swap!(x::ADT, bond::Int; trunc::TruncationScheme=DefaultTruncation)
+function swap!(x::ADT, bond::Int; trunc::TruncationScheme=DefaultITruncation)
 	x[bond], x.s[bond+1], x[bond+1] = _swap_gate(x.s[bond], x[bond], x.s[bond+1], x[bond+1], trunc=trunc)
 	return x
 end